backend/requirements.txt
Flask==3.0.3
Flask-Cors
GitPython==3.1.43
google-cloud-build==3.23.0
google-cloud-run
google-cloud-storage
gunicorn
python-dotenv
langchain
langchain-core
langchain-google-genai
PyGithub
flask-swagger-ui
requests
